requirements.txt
aniso8601==9.0.1
blinker==1.7.0
boto==2.49.0
boto3==1.34.23
botocore==1.34.23
charset-normalizer==3.2.0
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.1
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.4
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
requests